refactor(header): rename HeaderButtons to HeaderNav - update component for better semantics - update all imports in related files
