+++ src/components/Header/components/index.ts
@@ -1,2 +1,2 @@
-export { default as HeaderButtons } from './HeaderButtons/HeaderButtons';
+export { default as HeaderNav } from './HeaderNav/HeaderNav';
 export { default as HeaderSearch } from './HeaderSearch/HeaderSearch';
